requirements.txt
PyQt6>=6.2.0
Pillow>=9.0.0
numpy>=1.21.0
exifread>=2.3.2
pywin32>=304; platform_system == "Windows"
